Add and use localStorageInitialValues in place of hard-coded values in useLocalStorage.
